[IPC] Fix Misra C issues
- Following MISRA C issues are fixed
    - Addition of unsigned and signed values
    - Fix the data type for ctrl task

+++ packages/ti/drv/ipc/ipc.h
@@ -169,7 +169,7 @@ typedef struct RPMessage_Params_s
     uint32_t  stackSize;
     /**< StackSize used for the task */
 
-    uint32_t ctrlTaskPriority;
+    int8_t ctrlTaskPriority;
     /* Priority of ctrl Task */
 } RPMessage_Params;
